requirements.txt
awscli==1.32.7
blessed==1.20.0
botocore==1.34.7
chalice==1.29.0
click==8.1.7
colorama==0.4.4
coverage==7.3.4
docutils==0.16
iniconfig==2.0.0
inquirer==2.10.1
jmespath==1.0.1
packaging==23.2
pdfkit==1.0.0
Pillow==10.1.0
pluggy==1.3.0
pyasn1==0.5.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.0.5
rsa==4.7.2
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.12
